# api.yaml
openapi: 3.0.0
info:
  title: Aspire APIs
  description: For the enterprise use
  version: 1.0.0
  contact:
    email: soham.patel@aspiresoftserv.in
servers:
  - url: http://localhost:3000
    description: Internal server for testing
paths:
...
  /users/{userId}:
    get:
      summary: Returns a user by ID.
      tags:
      - User API
      parameters:
        - name: userId
          in: path
          required: true
          description: The ID of the user to return.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: A user object.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 12
        "400":
          description: The specified user ID is invalid (not a number).
        "404":
          description: A user with the specified ID was not found.
        default:
          description: Unexpected error
    patch:
      summary: Update a user.
      tags:
      - User API
      responses:
        "200":
          description: Created

    delete:
      summary: Delete a user.
      tags:
      - User API
      responses:
        "200":
          description: Created